Adjust tearDown() to deal with skipped tests (eliminates "PHP Fatal error:  Call to a member function getObjects() on a non-object").

+++ tests/StreamTest.php
@@ -81,6 +81,9 @@ class Services_Amazon_S3_StreamTest extends PHPUnit_Framework_TestCase
      */
     protected function tearDown()
     {
+        if (empty($this->bucket)) {
+            return;
+        }
         foreach ($this->bucket->getObjects() as $object) {
             $object->delete();
         }
