+++ tests/Test.php
@@ -66,6 +66,9 @@ class Services_Amazon_S3_Test extends PHPUnit_Framework_TestCase
      */
     protected function tearDown()
     {
+        if (empty($this->bucket)) {
+            return;
+        }
         foreach ($this->bucket->getObjects() as $object) {
             $object->delete();
         }
